added src/test (#40)
I added a src/test folder and moved the tests in there so that they can be in the same package as the source code they're testing. I wrote some tests in TestBanker and completed the TestSuite. I also added some questions in comments in Turn because I think that class still needs some work