Changes to turn (#47)
* added src/test

I added a src/test folder and moved the tests in there so that they can be in the same package as the source code they're testing. I wrote some tests in TestBanker and completed the TestSuite. I also added some questions in comments in Turn because I think that class still needs some work

* added more method headers

also deleted the TestSquare class

* changes didn't commit

* blah

* Turn changes

I combined rollDice and movePiece (from Board) into a takeTurn method because I felt the Board was overstepping its bounds to call performAction. The method "rollDice" now does a lot more than rolling dice, hence the new name. I also set up takeTurn to call notifyObservers with the GUI being the intended Observer. Finally I added a currentIndex variable to Player. This variable will allow Turn to keep track of where the player is in order to index into Board's Square array and call performAction on the Square on which the player is now located.

* playGame

implemented playGame in Monopoly

* update() in GUI

made update method in GUI and implemented Observer